requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.2.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
easyocr==1.7.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.4
npx==0.1.6
numpy==2.0.2
ollama==0.4.8
opencv-python-headless==4.11.0.86
orjson==3.10.16
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pyasn1==0.4.8
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyPDF2==3.0.1
python-bidi==0.6.6
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.28.1
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.13
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
